requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.9
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
pika==1.3.1
protobuf==3.20.1
PyJWT==2.6.0
pytz==2022.6
six==1.16.0
Werkzeug==2.2.2
zipp==3.10.0
